requirements.txt
python-telegram-bot
pandas
pandas_ta
pytesseract
opencv-python-headless
openai
pytz
python-dotenv